fix: improve federated module state management
Previously we saw a the dynamic script get reloaded if we had nested federated modules. By moving the dynamic script loading into a dedicated component and not loading that component until the module info is ready we avoid this problem.

Also refactor the code into a number of files and add missing declarations / types
Also cache the terms review state in a ref