fix: make the local development workflow work again
console.redhat.com is now being served with a CSP header that disables any javascript coming from sources not listed in the policy.
We load the federated modules from localhost (many ports) when running the local dev environment, the CSP breaks this behavior.

To fix it, we need to make the app ask for the resources through a pre-defined path, and then proxy the request to the right local server.
To do this the config.json contains a new basePath property that specifies the pre-defined path mentioned before, and a new proxyTarget property
that contains where the proxy should point to fetch the resources for all requests that contains the pre-defined path.