fix: reorder routes from less specific to more specific ones
This fixes the case where some deep routes have something in common with a less deep url.

Say we have the two following exact routes:
/overview
/streams/overview

If the user navigates to /streams/overview, both routes will match, but only the last one will be triggered by the wrapping Switch. If we were to reverse the order, it would be the /overview route to be triggered which we don't want.

So it's important to always sort the routes by specificity when adding them to the list of routes.